fix decimals

+++ spark_fhir_schemas/r4/complex_types/codesystem_property1.py
@@ -5,6 +5,7 @@ from typing import Union
 from pyspark.sql.types import ArrayType
 from pyspark.sql.types import BooleanType
 from pyspark.sql.types import DataType
+from pyspark.sql.types import FloatType
 from pyspark.sql.types import IntegerType
 from pyspark.sql.types import StringType
 from pyspark.sql.types import StructField
@@ -125,7 +126,7 @@ class CodeSystem_Property1Schema:
                 # The value of this property.
                 StructField("valueDateTime", StringType(), True),
                 # The value of this property.
-                StructField("valueDecimal", IntegerType(), True),
+                StructField("valueDecimal", FloatType(), True),
             ]
         )
         if not include_extension:
